refactor: 优化代码细节并统一类型别名
1. 替换字符串拼接的\n为字符字面量'\n'
2. 使用from_ref简化切片创建，遍历index entries时直接取keys
3. 重构条件判断合并逻辑，简化重复赋值代码
4. 新增ObjectList类型别名，统一返回类型
5. 优化Tls流存储为Box指针，简化packfile解析循环逻辑
6. 修复zlib校验判断写法，优化url仓库名提取方法
7. 移除冗余的类型转换注释和代码

+++ src/commands/push.rs
@@ -97,10 +97,10 @@ fn generate_pack(objects: &[(String, Vec<u8>)]) -> Result<Vec<u8>, Box<dyn std::
         let size = raw_content.len();
         let mut size_bytes = Vec::new();
         let mut remaining = size;
-        size_bytes.push(((type_code << 4) | (remaining as u8 & 0x0F)) as u8);
+        size_bytes.push((type_code << 4) | (remaining as u8 & 0x0F));
         remaining >>= 4;
         while remaining > 0 {
-            size_bytes.push((0x80 | (remaining as u8 & 0x7F)) as u8);
+            size_bytes.push(0x80 | (remaining as u8 & 0x7F));
             remaining >>= 7;
         }
         if size_bytes.len() > 1 {
